requirements.txt
# CARLA Client
carla==0.9.15

# YOLO and Computer Vision
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Utilities
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Visualization and Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.1.0

# VNC/Remote Access (for container)
websockify>=0.11.0

# Logging and Monitoring
loguru>=0.7.0
psutil>=5.9.0
